add project-card.component and functionalitiy
